+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - fix: step request elapsed timing should contain ContentTransfer part
 - fix #1288: unable to go get httprunner v4
+- feat: support converting Postman collection to HttpRunner testcase
 
 **python version**